my implementation of the assembler of unit 6, in c.
I've used the suggested pattern and workflow recommended by noam and shimon in this unit.
i did it in c language, it may be seen a bit long and complicated but I've made sure to include as many explanations in the comments as possible to provide context and connection to other files.
notice:  it seems that instead of creating xxx.hack files it just created xxxhack. its a stupid bug but for now ill upload this version.